+++ packages/frontend/src/features/scheduler/components/SchedulersListItem.tsx
@@ -1,16 +1,12 @@
-import {
-    Button,
-    Classes,
-    Menu,
-    MenuDivider,
-    PopoverPosition,
-} from '@blueprintjs/core';
-import { MenuItem2, Popover2 } from '@blueprintjs/popover2';
+import { Classes } from '@blueprintjs/core';
 import {
     getHumanReadableCronExpression,
     SchedulerAndTargets,
 } from '@lightdash/common';
+import { Divider, Menu } from '@mantine/core';
+import { IconDots, IconPencil, IconTrash } from '@tabler/icons-react';
 import { FC } from 'react';
+import MantineIcon from '../../../components/common/MantineIcon';
 import {
     InfoContainer,
     PageDetailsContainer,
@@ -40,30 +36,29 @@ const SchedulersListItem: FC<SchedulersListItemProps> = ({
                 className={Classes.TEXT_OVERFLOW_ELLIPSIS}
             >
                 <SchedulerName>{scheduler.name}</SchedulerName>
-                <Popover2
-                    content={
-                        <Menu>
-                            <MenuItem2
-                                icon="edit"
-                                text="Edit"
-                                onClick={() => onEdit(scheduler.schedulerUuid)}
-                            />
-                            <MenuDivider />
-                            <MenuItem2
-                                icon="delete"
-                                intent="danger"
-                                text="Delete"
-                                onClick={() =>
-                                    onDelete(scheduler.schedulerUuid)
-                                }
-                            />
-                        </Menu>
-                    }
-                    position={PopoverPosition.BOTTOM_RIGHT}
-                    lazy
-                >
-                    <Button minimal icon="more" />
-                </Popover2>
+                <Menu withArrow withinPortal width={100}>
+                    <Menu.Target>
+                        <MantineIcon icon={IconDots} />
+                    </Menu.Target>
+
+                    <Menu.Dropdown>
+                        <Menu.Item
+                            icon={<MantineIcon icon={IconPencil} />}
+                            onClick={() => onEdit(scheduler.schedulerUuid)}
+                        >
+                            Edit
+                        </Menu.Item>
+                        <Divider />
+
+                        <Menu.Item
+                            icon={<MantineIcon color="red" icon={IconTrash} />}
+                            onClick={() => onDelete(scheduler.schedulerUuid)}
+                            color="red"
+                        >
+                            Delete
+                        </Menu.Item>
+                    </Menu.Dropdown>
+                </Menu>
             </SchedulerDetailsContainer>
             <PageDetailsContainer>
                 <UpdatedInfoLabel>